break up connection.js
