+++ src/directedConnection.js
@@ -1,0 +1,18 @@
+import { Connection } from './connection.js'
+
+// for directed connections between two jurisdictions
+export class DirectedConnection extends Connection{
+	#to
+	#from
+	constructor(source,target){
+		super(source,target)
+		this.#from = source
+		this.#to = target
+	}
+	get from(){
+		return this.#from
+	}
+	get to(){
+		return this.#to
+	}
+}
